+++ commands/devTools/shutdown.js
@@ -4,6 +4,8 @@ const strings = require('../../res/strings');
 
 const uidR = process.env.UIDR;
 const uidJ = process.env.UIDJ;
+const uidD = process.env.UIDD;
+const uidT = process.env.UIDT;
 
 const { warnUser } = require('../../functions/warnUser.js');
 
@@ -15,7 +17,7 @@ module.exports = {
 	uses: strings.COMMAND_USES_DEVS,
 	syntax: `${prefix}shutdown`,
 	async execute(client, message, args) {
-		if (message.author.id === uidR || message.author.id === uidJ) {
+		if (message.author.id === uidR || message.author.id === uidJ || message.author.id === uidD || message.author.id === uidT) {
 			await message.inlineReply('Shutting down...');
 			await process.exit();
 		}