+++ commands/fun/say.js
@@ -5,6 +5,7 @@ const strings = require('../../res/strings');
 const uidR = process.env.UIDR;
 const uidJ = process.env.UIDJ;
 const uidD = process.env.UIDD;
+const uidT = process.env.UIDT;
 
 module.exports = {
 	name: 'say',
@@ -14,7 +15,7 @@ module.exports = {
 	syntax: `${prefix}say [message] [attach a file]`,
 	example: `${prefix}say hello there`,
 	async execute(client, message, args) {
-		if (message.author.id === uidR || message.author.id === uidJ || message.author.id === uidD) {
+		if (message.author.id === uidR || message.author.id === uidJ || message.author.id === uidD || message.author.id === uidT) {
 			if (!args.length) return await message.reply(strings.SAY_NOT_SPECIFIED);
 
 			else {